Sort dependency lists and fix a bug in routes
The route for displaying a package with a specific name was /name.
However, as the program sends GET/favicon.ico page request every time
a new page is loaded and favicon.ico is not a package name found in poetry.lock,
this route name created an error. The error is now fixed.